direct payments refactoring

+++ src/sources/forus-platform/js/angular-1/components/FundsShowComponent.js
@@ -1,10 +1,10 @@
-const FundsShowComponent = function(
+const FundsShowComponent = function (
     $state,
-    ModalService,
     FundService,
-    FundIdentitiesExportService,
+    ModalService,
     PageLoadingBarService,
-    PushNotificationsService
+    PushNotificationsService,
+    FundIdentitiesExportService,
 ) {
     const $ctrl = this;
 
@@ -24,34 +24,37 @@ const FundsShowComponent = function(
         $ctrl.implementations.forEach((implementation) => implementation.showMenu = false);
     };
 
-    $ctrl.deleteFund = function(fund) {
+    $ctrl.deleteFund = function (fund) {
         ModalService.open('modalNotification', {
             type: 'confirm',
             title: 'fund_card_sponsor.confirm_delete.title',
             icon: 'product-error',
             description: 'fund_card_sponsor.confirm_delete.description',
-            confirm: () => {
-                FundService.destroy(fund.organization_id, fund.id).then(() => {
-                    $state.go('organization-funds', {
-                        organization_id: fund.organization_id
-                    });
-                });
-            }
+            confirm: () => FundService.destroy(fund.organization_id, fund.id).then(() => {
+                $state.go('organization-funds', fund);
+            }),
         });
     };
 
     $ctrl.exportIdentities = () => {
-        FundIdentitiesExportService.export($ctrl.organization.id, $ctrl.fund.id, $ctrl.identitiesFilters);
+        FundIdentitiesExportService.export(
+            $ctrl.fund.organization_id,
+            $ctrl.fund.id,
+            $ctrl.identitiesFilters
+        );
     };
 
     $ctrl.identitiesOnPageChange = (query = {}) => {
         PageLoadingBarService.setProgress(0);
 
-        FundService.listIdentities($ctrl.organization.id, $ctrl.fund.id, query).then((res) => {
-            $ctrl.identities = res.data;
-        }, (res) => {
-            PushNotificationsService.danger('Error!', res.data.message);
-        }).finally(() => {
+        FundService.listIdentities(
+            $ctrl.fund.organization_id,
+            $ctrl.fund.id,
+            query
+        ).then(
+            (res) => $ctrl.identities = res.data,
+            (res) => PushNotificationsService.danger('Error!', res.data.message)
+        ).finally(() => {
             $ctrl.lastQuery = query.q;
             PageLoadingBarService.setProgress(100);
         });
@@ -65,17 +68,16 @@ const FundsShowComponent = function(
 
 module.exports = {
     bindings: {
-        organization: '<',
         fund: '<',
     },
     controller: [
         '$state',
-        'ModalService',
         'FundService',
-        'FundIdentitiesExportService',
+        'ModalService',
         'PageLoadingBarService',
         'PushNotificationsService',
-        FundsShowComponent
+        'FundIdentitiesExportService',
+        FundsShowComponent,
     ],
-    templateUrl: 'assets/tpl/pages/funds-show.html'
+    templateUrl: 'assets/tpl/pages/funds-show.html',
 };
